+++ emanote/src/Emanote/View/Template.hs
@@ -2,6 +2,7 @@ module Emanote.View.Template (emanoteSiteOutput, render) where
 
 import Control.Monad.Logger (MonadLoggerIO)
 import Data.Aeson.Types qualified as Aeson
+import Data.Map.Strict qualified as Map
 import Data.Map.Syntax ((##))
 import Data.Set qualified as Set
 import Data.Text qualified as T
@@ -70,24 +71,26 @@ render :: (MonadIO m, MonadLoggerIO m) => Model -> SR.SiteRoute -> m (Ema.Asset 
 render m = \case
   SR.SiteRoute_MissingR urlPath -> do
     let hereRoute = R.decodeHtmlRoute urlPath
+        (_, meta) = C.defaultRouteMeta m
         note404 =
           MN.missingNote hereRoute (toText urlPath)
-            & setErrorPageMeta
+            & setErrorPageMeta meta
             & MN.noteTitle
-            .~ "! Missing link"
+            .~ fromString (toString $ C.i18nText meta "missingLink" "! Missing link")
     pure $ Ema.AssetGenerated Ema.Html $ renderLmlHtml m note404
   SR.SiteRoute_AmbiguousR urlPath notes -> do
-    let noteAmb =
+    let (_, meta) = C.defaultRouteMeta m
+        noteAmb =
           MN.ambiguousNoteURL urlPath notes
-            & setErrorPageMeta
+            & setErrorPageMeta meta
             & MN.noteTitle
-            .~ "! Ambiguous link"
+            .~ fromString (toString $ C.i18nText meta "ambiguousLink" "! Ambiguous link")
     pure $ Ema.AssetGenerated Ema.Html $ renderLmlHtml m noteAmb
   SR.SiteRoute_ResourceRoute r -> pure $ renderResourceRoute m r
   SR.SiteRoute_VirtualRoute r -> renderVirtualRoute m r
   where
-    setErrorPageMeta =
-      MN.noteMeta .~ SData.mergeAesons (withTemplateName "/templates/error" :| [withSiteTitle "Emanote Error"])
+    setErrorPageMeta meta =
+      MN.noteMeta .~ SData.mergeAesons (withTemplateName "/templates/error" :| [withSiteTitle (C.i18nText meta "emanoteError" "Emanote Error")])
 
 renderResourceRoute :: Model -> SR.ResourceRoute -> Ema.Asset LByteString
 renderResourceRoute m = \case
@@ -124,8 +127,8 @@ renderSRIndex model = do
   let (r, meta) = C.defaultRouteMeta model
       tCtx = C.mkTemplateRenderCtx model r meta
   C.renderModelTemplate model "templates/special/index" $ do
-    C.commonSplices ($ emptyRenderCtx) model meta "Index"
-    routeTreeSplices tCtx Nothing model
+    C.commonSplices ($ emptyRenderCtx) model meta (fromString . toString $ C.i18nText meta "index" "Index")
+    routeTreeSplices tCtx meta Nothing model
 
 loaderHead :: LByteString
 loaderHead =
@@ -151,7 +154,7 @@ renderLmlHtml :: Model -> MN.Note -> LByteString
 renderLmlHtml model note = do
   let r = note ^. MN.noteRoute
       meta = patchMeta $ Meta.getEffectiveRouteMetaWith (note ^. MN.noteMeta) r model
-      doc = prependDataErrors (Meta.cascadeYamlErrors model r) (note ^. MN.noteDoc)
+      doc = prependDataErrors meta (Meta.cascadeYamlErrors model r) (note ^. MN.noteDoc)
       toc = newToc doc
       sourcePath = fromMaybe (R.withLmlRoute R.encodeRoute r) $ do
         fmap snd $ note ^. MN.noteSource
@@ -171,7 +174,7 @@ renderLmlHtml model note = do
           hasFlag = if flag == "toc" then hasFlag' && not (tocUnnecessaryToRender toc) else hasFlag'
       "ema:has:" <> flag ## Heist.ifElseISplice hasFlag
     -- Sidebar navigation
-    routeTreeSplices ctx (Just r) model
+    routeTreeSplices ctx meta (Just r) model
     "ema:breadcrumbs" ##
       C.routeBreadcrumbs ctx model r
     -- Note stuff
@@ -257,8 +260,8 @@ dayIsoText =
 
 If there is no 'current route', all sub-trees are marked as active/open.
 -}
-routeTreeSplices :: (Monad n) => C.TemplateRenderCtx n -> Maybe R.LMLRoute -> Model -> H.Splices (HI.Splice Identity)
-routeTreeSplices tCtx mCurrentRoute model = do
+routeTreeSplices :: (Monad n) => C.TemplateRenderCtx n -> Aeson.Value -> Maybe R.LMLRoute -> Model -> H.Splices (HI.Splice Identity)
+routeTreeSplices tCtx meta mCurrentRoute model = do
   "ema:route-tree" ##
     Splices.treeSplice getOrder (model ^. M.modelFolgezettelTree)
       $ \(last -> nodeRoute) children -> do
@@ -285,7 +288,16 @@ routeTreeSplices tCtx mCurrentRoute model = do
         "node:active" ## Heist.ifElseISplice isActiveNode
         "node:activeTree" ## Heist.ifElseISplice isActiveTree
         "node:terminal" ## Heist.ifElseISplice (null children)
-        "tree:childrenCount" ## HI.textSplice (show $ length children)
+        let childrenCount = show $ length children
+        "tree:childrenCount" ## HI.textSplice childrenCount
+        "tree:childrenInsideTitle" ##
+          HI.textSplice
+            ( C.i18nTextWith
+                meta
+                "childrenInsideTitle"
+                (Map.fromList [("count", childrenCount)])
+                (childrenCount <> " children inside")
+            )
         "tree:open" ## Heist.ifElseISplice openTree
         "has-current-route" ## Heist.ifElseISplice (isJust mCurrentRoute)
   where
@@ -323,7 +335,7 @@ Stays at render time so the banner can be scoped to this note's cascade
 affected note. Banner construction goes through 'MN.errorDiv' so the
 markdown and yaml error surfaces share one Div shape. Issue #285.
 -}
-prependDataErrors :: [Text] -> Pandoc -> Pandoc
-prependDataErrors [] doc = doc
-prependDataErrors errs (Pandoc m blocks) =
-  Pandoc m (MN.errorDiv "Emanote: bad YAML files" errs : blocks)
+prependDataErrors :: Aeson.Value -> [Text] -> Pandoc -> Pandoc
+prependDataErrors _ [] doc = doc
+prependDataErrors meta errs (Pandoc m blocks) =
+  Pandoc m (MN.errorDiv (C.i18nText meta "badYamlFiles" "Emanote: bad YAML files") errs : blocks)